myriad: refactor out of parallel into base package
DbException
Query
