+++ src/edu/washington/escience/myriad/operator/DbIterator.java
@@ -3,8 +3,8 @@ package edu.washington.escience.myriad.operator;
 import java.io.Serializable;
 import java.util.NoSuchElementException;
 
+import edu.washington.escience.myriad.DbException;
 import edu.washington.escience.myriad.Schema;
-import edu.washington.escience.myriad.parallel.DbException;
 import edu.washington.escience.myriad.table._TupleBatch;
 
 /**
