+++ src/edu/washington/escience/myriad/operator/JdbcQueryScan.java
@@ -2,10 +2,10 @@ package edu.washington.escience.myriad.operator;
 
 import java.util.Iterator;
 
+import edu.washington.escience.myriad.DbException;
 import edu.washington.escience.myriad.Schema;
 import edu.washington.escience.myriad.TupleBatch;
 import edu.washington.escience.myriad.accessmethod.JdbcAccessMethod;
-import edu.washington.escience.myriad.parallel.DbException;
 import edu.washington.escience.myriad.table._TupleBatch;
 
 public class JdbcQueryScan extends Operator {
