+++ src/edu/washington/escience/myriad/operator/SQLiteQueryScan.java
@@ -2,10 +2,10 @@ package edu.washington.escience.myriad.operator;
 
 import java.util.Iterator;
 
+import edu.washington.escience.myriad.DbException;
 import edu.washington.escience.myriad.Schema;
 import edu.washington.escience.myriad.TupleBatch;
 import edu.washington.escience.myriad.accessmethod.SQLiteAccessMethod;
-import edu.washington.escience.myriad.parallel.DbException;
 import edu.washington.escience.myriad.table._TupleBatch;
 
 public class SQLiteQueryScan extends Operator {
