+++ src/edu/washington/escience/myriad/operator/SQLiteSQLProcessor.java
@@ -1,7 +1,7 @@
 package edu.washington.escience.myriad.operator;
 
+import edu.washington.escience.myriad.DbException;
 import edu.washington.escience.myriad.Schema;
-import edu.washington.escience.myriad.parallel.DbException;
 import edu.washington.escience.myriad.table._TupleBatch;
 
 public class SQLiteSQLProcessor extends SQLiteQueryScan {
