+++ src/edu/washington/escience/myriad/parallel/BlockingDataReceiver.java
@@ -1,5 +1,6 @@
 package edu.washington.escience.myriad.parallel;
 
+import edu.washington.escience.myriad.DbException;
 import edu.washington.escience.myriad.Schema;
 import edu.washington.escience.myriad.operator.Operator;
 import edu.washington.escience.myriad.table._TupleBatch;
