+++ src/edu/washington/escience/myriad/parallel/ShuffleConsumer.java
@@ -3,6 +3,7 @@ package edu.washington.escience.myriad.parallel;
 import java.util.BitSet;
 import java.util.HashMap;
 
+import edu.washington.escience.myriad.DbException;
 import edu.washington.escience.myriad.Schema;
 import edu.washington.escience.myriad.operator.Operator;
 import edu.washington.escience.myriad.table._TupleBatch;
