+++ src/edu/washington/escience/myriad/table/DbAppendWriter.java
@@ -1,6 +1,6 @@
 package edu.washington.escience.myriad.table;
 
-import edu.washington.escience.myriad.parallel.DbException;
+import edu.washington.escience.myriad.DbException;
 
 public interface DbAppendWriter extends DbTable {
   public void append(_TupleBatch tb) throws DbException;
