+++ test/edu/washington/escience/myriad/accessmethod/JdbcAccessMethodTest.java
@@ -7,11 +7,11 @@ import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
 
+import edu.washington.escience.myriad.DbException;
 import edu.washington.escience.myriad.Schema;
 import edu.washington.escience.myriad.TupleBatch;
 import edu.washington.escience.myriad.Type;
 import edu.washington.escience.myriad.operator.JdbcQueryScan;
-import edu.washington.escience.myriad.parallel.DbException;
 
 /**
  * @author dhalperi
